feat: Display recent view in cards - EXO-80883 - eXIP-24
This change will implement cards views
